+++ genhub/beebase.py
@@ -136,13 +136,14 @@ class BeeBaseDB(genhub.genomedb.GenomeDB):
 
 def test_scaffolds_download():
     """BeeBase consortium scaffolds download"""
+    registry = genhub.registry.Registry()
 
-    label, config = genhub.conf.load_one('conf/hym/Emex.yml')
+    config = registry.genome('Emex')
     testurl = ('http://hymenopteragenome.org/beebase/sites/'
                'hymenopteragenome.org.beebase/files/data/consortium_data/'
                'Eufriesea_mexicana.v1.0.fa.gz')
     testpath = './Emex/Eufriesea_mexicana.v1.0.fa.gz'
-    emex_db = BeeBaseDB(label, config)
+    emex_db = BeeBaseDB('Emex', config)
     assert emex_db.gdnaurl == testurl, \
         'scaffold URL mismatch\n%s\n%s' % (emex_db.gdnaurl, testurl)
     assert emex_db.gdnapath == testpath, \
@@ -152,13 +153,14 @@ def test_scaffolds_download():
 
 def test_annot_download():
     """BeeBase consortium annotation download"""
+    registry = genhub.registry.Registry()
 
-    label, config = genhub.conf.load_one('conf/hym/Dnov.yml')
+    config = registry.genome('Dnov')
     testurl = ('http://hymenopteragenome.org/beebase/sites/'
                'hymenopteragenome.org.beebase/files/data/consortium_data/'
                'Dufourea_novaeangliae_v1.1.gff.gz')
     testpath = 'BeeBase/Dnov/Dufourea_novaeangliae_v1.1.gff.gz'
-    dnov_db = BeeBaseDB(label, config, workdir='BeeBase')
+    dnov_db = BeeBaseDB('Dnov', config, workdir='BeeBase')
     assert dnov_db.gff3url == testurl, \
         'annotation URL mismatch\n%s\n%s' % (dnov_db.gff3url, testurl)
     assert dnov_db.gff3path == testpath, \
@@ -167,13 +169,14 @@ def test_annot_download():
 
 def test_proteins_download():
     """BeeBase consortium protein download"""
+    registry = genhub.registry.Registry()
 
-    label, config = genhub.conf.load_one('conf/hym/Hlab.yml')
+    config = registry.genome('Hlab')
     testurl = ('http://hymenopteragenome.org/beebase/sites/'
                'hymenopteragenome.org.beebase/files/data/consortium_data/'
                'Habropoda_laboriosa_v1.2.pep.fa.gz')
     testpath = '/opt/db/genhub/Hlab/Habropoda_laboriosa_v1.2.pep.fa.gz'
-    hlab_db = BeeBaseDB(label, config, workdir='/opt/db/genhub')
+    hlab_db = BeeBaseDB('Hlab', config, workdir='/opt/db/genhub')
     assert hlab_db.proturl == testurl, \
         'protein URL mismatch\n%s\n%s' % (hlab_db.proturl, testurl)
     assert hlab_db.protpath == testpath, \
@@ -182,9 +185,10 @@ def test_proteins_download():
 
 def test_gdna_format():
     """BeeBase gDNA formatting"""
+    registry = genhub.registry.Registry()
 
-    label, conf = genhub.conf.load_one('conf/hym/Hlab.yml')
-    hlab_db = BeeBaseDB(label, conf, workdir='testdata/demo-workdir')
+    conf = registry.genome('Hlab')
+    hlab_db = BeeBaseDB('Hlab', conf, workdir='testdata/demo-workdir')
     hlab_db.preprocess_gdna(logstream=None, verify=False)
     outfile = 'testdata/demo-workdir/Hlab/Hlab.gdna.fa'
     testoutfile = 'testdata/fasta/hlab-first-6-out.fa'
@@ -193,9 +197,10 @@ def test_gdna_format():
 
 def test_annotation_beebase():
     """BeeBase annotation formatting"""
+    registry = genhub.registry.Registry()
 
-    label, conf = genhub.conf.load_one('conf/hym/Hlab.yml')
-    hlab_db = BeeBaseDB(label, conf, workdir='testdata/demo-workdir')
+    conf = registry.genome('Hlab')
+    hlab_db = BeeBaseDB('Hlab', conf, workdir='testdata/demo-workdir')
     hlab_db.preprocess_gff3(logstream=None, verify=False)
     outfile = 'testdata/demo-workdir/Hlab/Hlab.gff3'
     testfile = 'testdata/gff3/beebase-format-hlab.gff3'
@@ -204,9 +209,10 @@ def test_annotation_beebase():
 
 def test_proteins_beebase():
     """BeeBase protein formatting"""
+    registry = genhub.registry.Registry()
 
-    label, conf = genhub.conf.load_one('conf/hym/Hlab.yml')
-    hlab_db = BeeBaseDB(label, conf, workdir='testdata/demo-workdir')
+    conf = registry.genome('Hlab')
+    hlab_db = BeeBaseDB('Hlab', conf, workdir='testdata/demo-workdir')
     hlab_db.preprocess_prot(logstream=None, verify=False)
     outfile = 'testdata/demo-workdir/Hlab/Hlab.all.prot.fa'
     testoutfile = 'testdata/fasta/hlab-first-20-prot-out.fa'
@@ -215,9 +221,10 @@ def test_proteins_beebase():
 
 def test_protids():
     """BeeBase: extract protein IDs from GFF3"""
+    registry = genhub.registry.Registry()
 
-    label, conf = genhub.conf.load_one('conf/hym/Hlab.yml')
-    db = BeeBaseDB(label, conf)
+    conf = registry.genome('Hlab')
+    db = BeeBaseDB('Hlab', conf)
     protids = ['Hlab050%d' % x for x in range(62, 75)]
     infile = 'testdata/gff3/hlab-238.gff3'
     testids = list()
@@ -230,9 +237,10 @@ def test_protids():
 
 def test_protmap():
     """BeeBase: extract protein-->iLocus mapping from GFF3"""
+    registry = genhub.registry.Registry()
 
-    label, conf = genhub.conf.load_one('conf/hym/Hlab.yml')
-    db = BeeBaseDB(label, conf)
+    conf = registry.genome('Hlab')
+    db = BeeBaseDB('Hlab', conf)
     mapping = {'Hlab05074': 'HlabILC-653102', 'Hlab05071': 'HlabILC-653096',
                'Hlab05062': 'HlabILC-653079', 'Hlab05068': 'HlabILC-653090',
                'Hlab05072': 'HlabILC-653098', 'Hlab05070': 'HlabILC-653094',