+++ genhub/exons.py
@@ -125,8 +125,10 @@ def prepare(db, logstream=sys.stderr):  # pragma: no cover
 
 def test_coding_sequences():
     """Extract coding sequences"""
-    label, config = genhub.conf.load_one('conf/modorg/Atha.yml')
-    db = genhub.refseq.RefSeqDB(label, config, workdir='testdata/demo-workdir')
+    registry = genhub.registry.Registry()
+    config = registry.genome('Atha')
+    db = genhub.refseq.RefSeqDB('Atha', config,
+                                workdir='testdata/demo-workdir')
     cds_sequences(db, logstream=None)
 
     outfile = 'testdata/demo-workdir/Atha/Atha.cds.fa'
@@ -136,8 +138,10 @@ def test_coding_sequences():
 
 def test_exon_sequences():
     """Extract exon sequences"""
-    label, config = genhub.conf.load_one('conf/modorg/Atha.yml')
-    db = genhub.refseq.RefSeqDB(label, config, workdir='testdata/demo-workdir')
+    registry = genhub.registry.Registry()
+    config = registry.genome('Atha')
+    db = genhub.refseq.RefSeqDB('Atha', config,
+                                workdir='testdata/demo-workdir')
     exon_sequences(db, logstream=None)
 
     outfile = 'testdata/demo-workdir/Atha/Atha.exons.fa'
@@ -147,8 +151,10 @@ def test_exon_sequences():
 
 def test_intron_sequences():
     """Extract intron sequences"""
-    label, config = genhub.conf.load_one('conf/modorg/Atha.yml')
-    db = genhub.refseq.RefSeqDB(label, config, workdir='testdata/demo-workdir')
+    registry = genhub.registry.Registry()
+    config = registry.genome('Atha')
+    db = genhub.refseq.RefSeqDB('Atha', config,
+                                workdir='testdata/demo-workdir')
     intron_sequences(db, logstream=None)
 
     outfile = 'testdata/demo-workdir/Atha/Atha.introns.fa'